requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.2.0
python-multipart==0.0.20
pydantic==2.8.2
pydantic-settings==2.4.0
email-validator==2.2.0
